Reorg TS common tests
